refactor(swift): share field-metadata rendering across dump and interface
The SwiftInterface module split moved the model-driven printer
(`SwiftDeclarationPrinter`) off `SwiftDump`'s metadata-aware dumpers, so
the interface lost every `// Field offset:` / `// Type Layout:` /
expanded-offset / enum-layout comment on the MachOImage path — the
runtime-metadata info RuntimeViewer and `swift-section interface` rely on.

Extract that field/enum metadata rendering into a single
`FieldLayoutRenderer` in `SwiftDeclarationRendering` and have both the
`SwiftDump` dumpers (Struct/Class/Enum) and the `SwiftPrinting` printer
route through it, restoring the lost output with one source of truth.
Verified against the pre-split (main) output: dump is byte-identical and
the interface's metadata comments match exactly.

Also corrects two metadata resolutions to read their value-witness table
through the same reader the metadata was resolved with
(`valueWitnessTable(in:)` / `asMetadataWrapper(in:)`), which otherwise
segfaults when enum layout / dump metadata is enabled on a MachOImage.
