+++ Sources/SwiftDump/Protocols/TypedDumper.swift
@@ -402,382 +402,3 @@ package enum BoundDumpedTypeNameRenderer {
     }
 }
 
-extension TypedDumper {
-    // MARK: - Previous (pre-substitution) implementation, kept here for
-    //         reference while reading the new walker. Deleting it does not
-    //         change behavior — the new `expandedFieldOffsets(...)` plus
-    //         `walkNestedExpandedFieldOffsets(of:...)` cover every case
-    //         this one did and additionally handle specialized generics.
-    //
-    // ```swift
-    // extension TypedDumper {
-    //     @SemanticStringBuilder
-    //     func expandedFieldOffsets(for mangledTypeName: MangledName, baseOffset: Int, baseIndentation: Int, ancestors: [Bool], in machO: MachOImage?) -> SemanticString {
-    //         let metatype: Any.Type?
-    //         if let machO {
-    //             metatype = try? RuntimeFunctions.getTypeByMangledNameInContext(mangledTypeName, in: machO)
-    //         } else {
-    //             metatype = try? RuntimeFunctions.getTypeByMangledNameInContext(mangledTypeName)
-    //         }
-    //         if let metatype,
-    //            let metadata = try? Metadata.createInProcess(metatype).asMetadataWrapper().struct,
-    //            let descriptor = try? metadata.descriptor().struct, !descriptor.isGeneric,
-    //            let nestedFieldOffsets = try? metadata.fieldOffsets(for: descriptor),
-    //            let nestedFieldRecords = try? descriptor.fieldDescriptor().records() {
-    //             let fieldEntries = Array(zip(nestedFieldRecords, nestedFieldOffsets))
-    //             for (fieldIndex, (nestedFieldRecord, nestedRelativeOffset)) in fieldEntries.enumerated() {
-    //                 if let fieldName = try? nestedFieldRecord.fieldName() {
-    //                     let absoluteOffset = baseOffset + Int(nestedRelativeOffset)
-    //                     let isLastField = fieldIndex == fieldEntries.count - 1
-    //                     let nestedMangledTypeName = try? nestedFieldRecord.mangledTypeName()
-    //                     let typeName = nestedMangledTypeName.flatMap { try? MetadataReader.demangleType(for: $0).printSemantic(using: .default).string } ?? ""
-    //                     configuration.expandedFieldOffsetComment(fieldName: fieldName, typeName: typeName, offset: absoluteOffset, baseIndentation: baseIndentation, ancestors: ancestors, isLast: isLastField)
-    //
-    //                     if let nestedMangledTypeName {
-    //                         expandedFieldOffsets(for: nestedMangledTypeName, baseOffset: absoluteOffset, baseIndentation: baseIndentation, ancestors: ancestors + [isLastField], in: nil)
-    //                     }
-    //                 }
-    //             }
-    //         }
-    //     }
-    // }
-    // ```
-    //
-    // Diff against the new implementation, point-by-point:
-    //
-    //   1. Resolution at the *top* hop ignored `metadataContext` —
-    //      `getTypeByMangledNameInContext(mangledTypeName, in: machO)` cannot
-    //      substitute generic parameters. For a specialized
-    //      `Box<Int>` dumper, a field `let inner: SingleParameterBox<A>`
-    //      mangled name still references `A`; resolution returned nil and
-    //      the whole expansion was skipped. The new entry calls
-    //      `resolveFieldMetatype(for:in:)` first so the dumper's specialized
-    //      metadata performs the substitution.
-    //
-    //   2. The `!descriptor.isGeneric` guard rejected *every* generic
-    //      descriptor — including the legitimately specialized
-    //      `SingleParameterBox<Int>` we now want to recurse into. The new
-    //      walker drops that guard because the metadata it holds is already
-    //      a specialized in-process metadata, so `fieldOffsets(for:)` and
-    //      friends are well-defined.
-    //
-    //   3. The recursive call passed `in: nil` (in-process resolution) but
-    //      again with no substitution context. Nested fields whose mangled
-    //      names referenced *their parent's* generic parameters could not
-    //      resolve. The new `walkNestedExpandedFieldOffsets(of:...)` threads
-    //      the just-resolved struct metadata as substitution context for
-    //      the next hop, mirroring how Swift's runtime walks
-    //      generic-arguments arrays through nested specializations.
-    //
-    //   4. `Metadata.createInProcess(metatype).asMetadataWrapper().struct`
-    //      acted as a kind check: it returned nil for non-struct metatypes
-    //      (class / enum / builtin / function …) so recursion never
-    //      entered them. The new walker performs the same kind check
-    //      inside `walkNestedExpandedFieldOffsets(of: Any.Type, ...)` —
-    //      its switch dispatches `.struct` / `.enum` / `.optional` into
-    //      dedicated walkers and falls through to a `default` no-op for
-    //      every other kind, so class / builtin / function metatypes
-    //      never reach an unsafe `StructMetadata.createInProcess`-style
-    //      cast. An earlier draft of this refactor *lost* that guard by
-    //      replacing the chain with a bare
-    //      `StructMetadata.createInProcess(metatype)` — the latter reads
-    //      16 bytes of the metadata blindly, so a class metatype
-    //      produced a misaligned `StructMetadata` whose
-    //      `structDescriptor()` then trapped on its internal
-    //      `descriptor().struct!` force-unwrap. `try?` does not catch a
-    //      forced-unwrap trap, hence the visible `Fatal error: Unexpectedly
-    //      found nil` you'd see when dumping a struct with a class field.
-
-    /// Top-level entry: walks the dumper-owned struct's nested struct
-    /// fields, emitting `expandedFieldOffsetComment` lines.
-    ///
-    /// Substitution flows in two phases:
-    ///   - Top hop: the field's mangled name comes from the dumper's
-    ///     descriptor; specialization (if any) lives in `metadataContext`.
-    ///     Use `resolveFieldMetatype` so generic field types like `let x: A`
-    ///     resolve to the bound concrete type.
-    ///   - Recursive hops: the just-resolved nested struct's metadata is
-    ///     itself a specialized in-process metadata, and any further nested
-    ///     mangled names that mention *its* generic parameters need to
-    ///     substitute against it. `walkNestedExpandedFieldOffsets(of:...)`
-    ///     threads that context through the recursion.
-    @SemanticStringBuilder
-    func expandedFieldOffsets(for mangledTypeName: MangledName, baseOffset: Int, baseIndentation: Int, ancestors: [Bool], in machO: MachOImage?) -> SemanticString {
-        let topMetatype: Any.Type?
-        if let machO {
-            topMetatype = resolveFieldMetatype(for: mangledTypeName, in: machO)
-                ?? (try? RuntimeFunctions.getTypeByMangledNameInContext(mangledTypeName, in: machO))
-        } else {
-            topMetatype = try? RuntimeFunctions.getTypeByMangledNameInContext(mangledTypeName)
-        }
-        if let topMetatype {
-            walkNestedExpandedFieldOffsets(of: topMetatype, baseOffset: baseOffset, baseIndentation: baseIndentation, ancestors: ancestors)
-        }
-    }
-
-    /// Dispatches recursive expansion by metadata kind. Structs expose their
-    /// stored fields directly; Optional and other enum wrappers expose payload
-    /// case records that can themselves carry specialized struct payloads.
-    @SemanticStringBuilder
-    private func walkNestedExpandedFieldOffsets(of metatype: Any.Type, baseOffset: Int, baseIndentation: Int, ancestors: [Bool], depth: Int = 0) -> SemanticString {
-        if depth >= nestedFieldOffsetExpansionDepthLimit {
-            emitNestedFieldOffsetDepthLimitWarning(for: metatype)
-        } else if let wrapper = try? Metadata.createInProcess(metatype).asMetadataWrapper() {
-            switch wrapper {
-            case .struct(let metadata):
-                walkNestedStructFieldOffsets(of: metadata, baseOffset: baseOffset, baseIndentation: baseIndentation, ancestors: ancestors, depth: depth)
-            case .enum(let metadata),
-                 .optional(let metadata):
-                walkNestedEnumPayloadFieldOffsets(of: metadata, baseOffset: baseOffset, baseIndentation: baseIndentation, ancestors: ancestors, depth: depth)
-            default:
-                SemanticString()
-            }
-        }
-    }
-
-    /// Plain (non-builder) helper so the result-builder body of
-    /// `walkNestedExpandedFieldOffsets` stays valid — the `#log` macro
-    /// expands to a `Void`-typed closure invocation, which the builder
-    /// accepts via `buildPartialBlock(first: Void)`, but keeping the
-    /// diagnostics out of the builder body avoids surprising callers
-    /// reading the walker.
-    private func emitNestedFieldOffsetDepthLimitWarning(for metatype: Any.Type) {
-        #log(.info, "walkNestedExpandedFieldOffsets reached nested field-offset depth limit \(nestedFieldOffsetExpansionDepthLimit, privacy: .public) — truncating expansion of \(metatype, privacy: .public)")
-    }
-
-    /// Recursive walk over a nested struct's fields. Mangled names read here
-    /// come from in-process descriptor memory; generic parameter references are
-    /// bound to the parent's concrete arguments statically (no runtime generic
-    /// context — see the static substitution helpers below).
-    @SemanticStringBuilder
-    private func walkNestedStructFieldOffsets(of metadata: StructMetadata, baseOffset: Int, baseIndentation: Int, ancestors: [Bool], depth: Int) -> SemanticString {
-        if let descriptor = try? metadata.structDescriptor(),
-           let nestedFieldOffsets = try? metadata.fieldOffsets(for: descriptor),
-           let nestedFieldRecords = try? descriptor.fieldDescriptor().records() {
-            let fieldEntries = Array(zip(nestedFieldRecords, nestedFieldOffsets))
-            for (fieldIndex, (nestedFieldRecord, nestedRelativeOffset)) in fieldEntries.enumerated() {
-                if let fieldName = try? nestedFieldRecord.fieldName() {
-                    let absoluteOffset = baseOffset + Int(nestedRelativeOffset)
-                    let isLastField = fieldIndex == fieldEntries.count - 1
-                    let nestedMangledTypeName = try? nestedFieldRecord.mangledTypeName()
-                    let typeName = nestedTypeName(for: nestedMangledTypeName, parentMetadata: metadata)
-                    configuration.expandedFieldOffsetComment(fieldName: fieldName, typeName: typeName, offset: absoluteOffset, baseIndentation: baseIndentation, ancestors: ancestors, isLast: isLastField)
-
-                    if let nestedMangledTypeName,
-                       let resolvedMetatype = resolveNestedMetatype(for: nestedMangledTypeName, parentMetadata: metadata) {
-                        walkNestedExpandedFieldOffsets(of: resolvedMetatype, baseOffset: absoluteOffset, baseIndentation: baseIndentation, ancestors: ancestors + [isLastField], depth: depth + 1)
-                    }
-                }
-            }
-        }
-    }
-
-    /// Recursive walk over payload cases for Optional and enum wrappers.
-    /// Payloads all begin at the enum payload area, offset 0 relative to the
-    /// enum value, so the child offset starts at `baseOffset`.
-    @SemanticStringBuilder
-    private func walkNestedEnumPayloadFieldOffsets(of metadata: EnumMetadata, baseOffset: Int, baseIndentation: Int, ancestors: [Bool], depth: Int) -> SemanticString {
-        if let descriptor = try? metadata.enumDescriptor(),
-           descriptor.hasPayloadCases,
-           let records = try? descriptor.fieldDescriptor().records() {
-            let payloadRecords = Array(records.prefix(descriptor.numberOfPayloadCases))
-            for (payloadIndex, payloadRecord) in payloadRecords.enumerated() {
-                if let mangledTypeName = try? payloadRecord.mangledTypeName(),
-                   !mangledTypeName.isEmpty,
-                   let resolvedMetatype = resolveNestedMetatype(for: mangledTypeName, parentMetadata: metadata) {
-                    let fieldName = (try? payloadRecord.fieldName()) ?? "payload"
-                    let typeName = nestedTypeName(for: mangledTypeName, parentMetadata: metadata)
-                    let isLastPayload = payloadIndex == payloadRecords.count - 1
-                    configuration.expandedFieldOffsetComment(fieldName: fieldName, typeName: typeName, offset: baseOffset, baseIndentation: baseIndentation, ancestors: ancestors, isLast: isLastPayload)
-                    walkNestedExpandedFieldOffsets(of: resolvedMetatype, baseOffset: baseOffset, baseIndentation: baseIndentation, ancestors: ancestors + [isLastPayload], depth: depth + 1)
-                }
-            }
-        }
-    }
-
-    /// Resolves a nested field's mangled name to its concrete `Any.Type` for
-    /// continued offset-tree recursion.
-    ///
-    /// Two safe strategies, in order — neither ever calls the live runtime's
-    /// generic-substituting `specializedFrom:` overload (which walks the
-    /// descriptor's generic environment and PAC-faults inside injected
-    /// processes; see the substitution helpers below):
-    ///   1. Static binding: if the field's type IS one of the parent's generic
-    ///      parameters (e.g. `value: A`), read the already-instantiated
-    ///      concrete argument metadata straight out of the parent's specialized
-    ///      metadata — a fully-formed, runtime-built metadata pointer we only
-    ///      read, never re-authenticate. This lets generic fields keep
-    ///      expanding into their bound concrete type.
-    ///   2. Bare resolve: otherwise resolve the full mangled name with no
-    ///      generic context. Non-generic fields resolve; a compound type that
-    ///      still mentions an unbound parameter (e.g. `Bar<A>`) returns nil and
-    ///      simply stops expanding — offsets stay correct either way because
-    ///      they come from `metadata.fieldOffsets`, not from this call.
-    private func resolveNestedMetatype<ParentMetadata: ValueMetadataProtocol>(for mangledTypeName: MangledName, parentMetadata: ParentMetadata) -> Any.Type? {
-        // 1. Static binding for a field that IS one of the parent's generic
-        //    parameters (e.g. `value: A`).
-        if let boundType = staticallyBoundMetatype(for: mangledTypeName, parentMetadata: parentMetadata) {
-            return boundType
-        }
-        // 2. The context-free runtime resolver `fatalError`s (uncatchable —
-        //    `try?` cannot intercept it) inside `createDependentMemberType`
-        //    when the mangled name references a generic parameter or an
-        //    associated type but we pass no generic context to bind it. So
-        //    only hand the runtime a name we've confirmed is fully concrete;
-        //    anything that still mentions a dependent type stops expanding
-        //    rather than risk an abort. Offsets are unaffected (they come from
-        //    `metadata.fieldOffsets`).
-        guard let node = try? MetadataReader.demangleTypeUncached(for: mangledTypeName),
-              !nodeContainsDependentReference(node)
-        else { return nil }
-        return try? RuntimeFunctions.getTypeByMangledNameInContext(mangledTypeName)
-    }
-
-    /// Renders the human-readable type name for the `expandedFieldOffsetComment`
-    /// line, with the parent's generic parameters statically substituted by
-    /// their bound concrete types (so `value: A` prints as `value: Int`).
-    private func nestedTypeName<ParentMetadata: ValueMetadataProtocol>(for mangledTypeName: MangledName?, parentMetadata: ParentMetadata) -> String {
-        guard let mangledTypeName else { return "" }
-        if let substitutedNode = substitutedNestedTypeNode(for: mangledTypeName, parentMetadata: parentMetadata) {
-            return substitutedNode.printSemantic(using: .default).string
-        }
-        return (try? MetadataReader.demangleTypeUncached(for: mangledTypeName).printSemantic(using: .default).string) ?? ""
-    }
-
-    // MARK: - Static generic-argument substitution
-    //
-    // The expanded-field-offset walker substitutes a nested type's generic
-    // parameter placeholders with the concrete types bound in the parent's
-    // specialized in-process metadata WITHOUT calling
-    // `swift_getTypeByMangledNameInContext(specializedFrom:)`. That runtime
-    // entry computes the (depth,index)→argument mapping by walking the
-    // descriptor's parent chain and authenticating signed pointers, which
-    // PAC-faults on bound generics drawn from third-party frameworks across the
-    // ObjC boundary (e.g. `SwiftUI.ObservedObject<NSObjectSubclass>`) inside an
-    // injected process — an uncatchable hardware trap. Instead we reproduce the
-    // mapping statically: the flat index is pure arithmetic (`flatIndex = index`
-    // at depth 0, see `_depthIndexToFlatIndex` in the Swift runtime's
-    // MetadataLookup.cpp), and the argument array sits immediately after the
-    // metadata header — both readable without ever handing a synthesized
-    // generic context back to the runtime.
-    //
-    // Only depth-0 parameter references are bound (the case that covers every
-    // single-level generic container — Optional, Array, Set, Dictionary,
-    // ObservedObject, …). Deeper references (nested generic type definitions
-    // like `Outer<A>.Inner<B>`) are left unbound rather than risk a wrong
-    // substitution.
-
-    /// Demangles `mangledTypeName` and rewrites every depth-0 generic parameter
-    /// placeholder into the concrete type bound by `parentMetadata`. Returns the
-    /// rewritten node, or nil only when the name cannot be demangled at all.
-    private func substitutedNestedTypeNode<ParentMetadata: ValueMetadataProtocol>(for mangledTypeName: MangledName, parentMetadata: ParentMetadata) -> Node? {
-        guard let node = try? MetadataReader.demangleTypeUncached(for: mangledTypeName) else { return nil }
-        guard let keyArgumentFlags = topLevelGenericKeyArgumentFlags(of: parentMetadata) else {
-            return node
-        }
-        let keyArgumentCount = keyArgumentFlags.lazy.filter { $0 }.count
-        return substitutingGenericParameters(in: node, parentMetadata: parentMetadata, keyArgumentFlags: keyArgumentFlags, keyArgumentCount: keyArgumentCount)
-    }
-
-    /// If `mangledTypeName` demangles to exactly one of the parent's depth-0
-    /// generic parameters, returns the concrete argument metadata bound for it;
-    /// otherwise nil. Lets recursion descend through bound generic fields
-    /// without invoking the runtime substituter.
-    private func staticallyBoundMetatype<ParentMetadata: ValueMetadataProtocol>(for mangledTypeName: MangledName, parentMetadata: ParentMetadata) -> Any.Type? {
-        guard let node = try? MetadataReader.demangleTypeUncached(for: mangledTypeName) else { return nil }
-        let typeNode = innerTypeNode(of: node)
-        guard typeNode.kind == .dependentGenericParamType,
-              let (depthValue, indexValue) = genericParameterDepthAndIndex(of: typeNode),
-              depthValue == 0,
-              let keyArgumentFlags = topLevelGenericKeyArgumentFlags(of: parentMetadata),
-              let flatIndex = depthZeroFlatIndex(forIndex: indexValue, keyArgumentFlags: keyArgumentFlags)
-        else { return nil }
-        let keyArgumentCount = keyArgumentFlags.lazy.filter { $0 }.count
-        return boundGenericArgumentType(at: flatIndex, keyArgumentCount: keyArgumentCount, of: parentMetadata)
-    }
-
-    /// Recursively rewrites depth-0 `dependentGenericParamType` nodes into the
-    /// concrete bound argument's type node, leaving everything else intact.
-    private func substitutingGenericParameters<ParentMetadata: ValueMetadataProtocol>(in node: Node, parentMetadata: ParentMetadata, keyArgumentFlags: [Bool], keyArgumentCount: Int) -> Node {
-        if #available(macOS 11, iOS 14, tvOS 14, watchOS 7, *),
-           node.kind == .dependentGenericParamType,
-           let (depthValue, indexValue) = genericParameterDepthAndIndex(of: node),
-           depthValue == 0,
-           let flatIndex = depthZeroFlatIndex(forIndex: indexValue, keyArgumentFlags: keyArgumentFlags),
-           let argumentType = boundGenericArgumentType(at: flatIndex, keyArgumentCount: keyArgumentCount, of: parentMetadata),
-           let argumentMangledString = _mangledTypeName(argumentType),
-           let argumentNode = try? demangleAsNode(argumentMangledString, isType: true) {
-            return innerTypeNode(of: argumentNode)
-        }
-        let substitutedChildren = node.children.map {
-            substitutingGenericParameters(in: $0, parentMetadata: parentMetadata, keyArgumentFlags: keyArgumentFlags, keyArgumentCount: keyArgumentCount)
-        }
-        return Node.create(kind: node.kind, contents: node.contents, children: Array(substitutedChildren))
-    }
-
-    /// Reads the concrete generic argument metadata at `flatIndex` straight out
-    /// of a specialized in-process value metadata's inline argument array
-    /// (immediately after the metadata header). Returns nil if out of range or
-    /// the metadata is not in-process.
-    private func boundGenericArgumentType<ParentMetadata: ValueMetadataProtocol>(at flatIndex: Int, keyArgumentCount: Int, of parentMetadata: ParentMetadata) -> Any.Type? {
-        guard flatIndex >= 0, flatIndex < keyArgumentCount else { return nil }
-        guard let metadataPointer = try? parentMetadata.asPointer else { return nil }
-        let genericArgumentsBase = metadataPointer.advanced(by: MemoryLayout<ParentMetadata.Layout>.size)
-        let argumentBitPattern = genericArgumentsBase.load(fromByteOffset: flatIndex * MemoryLayout<UInt>.size, as: UInt.self)
-        guard argumentBitPattern != 0, let argumentPointer = UnsafeRawPointer(bitPattern: argumentBitPattern) else { return nil }
-        return unsafeBitCast(argumentPointer, to: Any.Type.self)
-    }
-
-    /// The parent value type's depth-0 generic parameters' `hasKeyArgument`
-    /// flags (in declaration order). nil when the parent is non-generic or its
-    /// generic context can't be read.
-    private func topLevelGenericKeyArgumentFlags<ParentMetadata: ValueMetadataProtocol>(of parentMetadata: ParentMetadata) -> [Bool]? {
-        guard let descriptor = try? parentMetadata.descriptor(),
-              let genericContext = try? descriptor.genericContext(),
-              let topLevelParameters = genericContext.allParameters.first
-        else { return nil }
-        return topLevelParameters.map(\.hasKeyArgument)
-    }
-
-    /// Maps a depth-0 (declaration-order) parameter index to its flat
-    /// key-argument index, skipping any non-key parameters before it. nil if the
-    /// index is out of range or the parameter itself carries no key argument.
-    private func depthZeroFlatIndex(forIndex index: Int, keyArgumentFlags: [Bool]) -> Int? {
-        guard index >= 0, index < keyArgumentFlags.count, keyArgumentFlags[index] else { return nil }
-        return keyArgumentFlags[0..<index].lazy.filter { $0 }.count
-    }
-
-    /// Extracts (depth, index) from a `dependentGenericParamType` node, whose
-    /// two `.index` children carry the depth and index respectively.
-    private func genericParameterDepthAndIndex(of node: Node) -> (depth: Int, index: Int)? {
-        let children = Array(node.children)
-        guard children.count == 2,
-              let depthValue = children[0].index,
-              let indexValue = children[1].index
-        else { return nil }
-        return (Int(depthValue), Int(indexValue))
-    }
-
-    /// Unwraps a `.type` wrapper node to its inner type node, leaving other
-    /// nodes unchanged. Splices a demangled concrete type into the position
-    /// previously held by a parameter placeholder.
-    private func innerTypeNode(of node: Node) -> Node {
-        if node.kind == .type, let firstChild = node.firstChild {
-            return firstChild
-        }
-        return node
-    }
-
-    /// True if the node tree references any unbound generic parameter or
-    /// associated type. These are exactly the cases that make the context-free
-    /// runtime resolver `fatalError` (uncatchable) instead of returning nil, so
-    /// callers must avoid handing such names to `swift_getTypeByMangledName`.
-    private func nodeContainsDependentReference(_ node: Node) -> Bool {
-        switch node.kind {
-        case .dependentGenericParamType, .dependentMemberType, .dependentAssociatedTypeRef:
-            return true
-        default:
-            return node.children.contains { nodeContainsDependentReference($0) }
-        }
-    }
-}
